+++ changelog.md
@@ -39,6 +39,8 @@
 
 # Unreleased
 
+- Create action ``actionbar``, allows to send a message in the action bar of the player
+
 # 1.0.3.1
 
 - Fix folia commands and data manager [#63](https://github.com/Maxlego08/zMenu/issues/63)